api/pyproject.toml
[tool.poetry]
name = "devloper-connect-api"
version = "1.0.0"
description = ""
authors = ["thorwolpert <thor@wolpert.ca>"]
license = "BSD 3-Clause"
readme = "README.md"
packages = [{include = "devloper-connect-api", from = "src"}]

[tool.poetry.dependencies]
python = "^3.12.4"
flask = "^1.1.4"
launchdarkly-server-sdk = "^8.3.0"
python-dotenv = "^1.0.1"
sentry-sdk = "^1.45.0"
flask-babel = "2.0.0"
flask-jwt-oidc = "^0.7.0"
gunicorn = "^22.0.0"
sbc-common-components = {git = "https://github.com/bcgov/sbc-common-components.git", rev = "master", subdirectory = "python"}
auth-api = {git = "https://github.com/bcgov/sbc-auth.git", rev = "c286045eb2cf9761cd57b59aa4c6e7b3146ea36c", subdirectory = "auth-api"}

[tool.poetry.group.test.dependencies]
freezegun = "^1.5.1"
hypothesis = "^6.108.0"
pexpect = "^4.9.0"
pytest = "^7.4.4"
pytest-cov = "^5.0.0"
pytest-mock = "^3.14.0"
requests-mock = "^1.12.1"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7.1"
faker = "^24.14.1"
lovely-pytest-docker = "^0.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
